pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easydl"
version = "1.1.0"
description = "A Python package for easy deep learning"
readme = "README.md"
authors = [
    {name = "Xiaonan Zhao", email = "zhaoxxn@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.7"
# core dependencies
dependencies = [
    "numpy>=1.19.0",
    "pillow>=8.0.0",
    "requests>=2.25.0",
    "pillow-heif>=0.10.0",
    "torch>=1.7.0",
    "torchvision>=0.17.2",
    "transformers>=4.0.0",
    "tqdm>=4.60.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "boto3>=1.38.0",
    "accelerate>=1.10.1",
]

[project.optional-dependencies]
# Dependencies for research
research = [
    "datasets>=4.4.1",
    "timm>=1.0.0",
    "plotly>=6.5.0",
    "nbformat>=5.10.4",
]

# Development tools
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
]

# Documentation tools
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

# All features combined
all = [
    "easydl[dev]",
    "easydl[research]",
    "easydl[docs]",
]

[project.urls]
Homepage = "https://github.com/xiaonanzzz/easydl"

[tool.setuptools.packages.find]
include = ["easydl*"]
exclude = ["tests*", "venvs*", "exp-ws*", "dev-setup*", "examples*"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3